backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pypdf==5.1.0
python-dotenv==1.0.1
transformers==4.46.0
torch==2.5.1
sentence-transformers==3.3.1
accelerate==1.2.1
